migrate a few additional tests and restore a preexisting one
